Improve detection of ChefFS-based commands in `knife rehash`
ChefFS-based commands have a superclass (Chef::ChefFS::Knife) which
defines its own inherited method that calls super. This breaks our
detection of where the subcommand is defined since the file with
the definition is no longer at the top of the call stack.

This commit special-cases subclasses with a superclass of
Chef::ChefFS::Knife to account for this.

Fixes #4089